Tune the PSLQ algorithm for no false positives, returning big.Int output
Convert to using big.Int for output, matrices etc.  This is because
the A matrix gets much bigger than the B matrix and was overflowing.

Use the A matrix to implement the final termination condition and tune
the exact ratio of precision so that no false positives are seen.

Factor the errors out into exported errors so we can test against them.

Raise minimum precision to 64 bits

Don't add extra precision on - it doesn't seem to be necessary

Add lots of tests to make sure we don't get false positives